+++ cryptostore/data/storage.py
@@ -1,5 +1,5 @@
 '''
-Copyright (C) 2018-2020  Bryant Moscon - bmoscon@gmail.com
+Copyright (C) 2018-2021  Bryant Moscon - bmoscon@gmail.com
 
 Please see the LICENSE file for the terms and conditions
 associated with this software.
@@ -13,17 +13,17 @@ from cryptostore.data.timescaledb import TimescaleDB
 
 
 class Storage(Store):
-    def __init__(self, config):
+    def __init__(self, config, **kwargs):
         self.config = config
         if isinstance(config.storage, list):
-            self.s = [Storage.__init_helper(s, config) for s in config.storage]
+            self.s = [Storage.__init_helper(s, config, **kwargs) for s in config.storage]
         else:
-            self.s = [Storage.__init_helper(config.storage, config)]
+            self.s = [Storage.__init_helper(config.storage, config, **kwargs)]
 
     @staticmethod
-    def __init_helper(store, config):
+    def __init_helper(store, config, **kwargs):
         if store == 'parquet':
-            return Parquet(config.parquet if 'parquet' in config else None)
+            return Parquet(config.exchanges, config.parquet if 'parquet' in config else None, **kwargs)
         elif store == 'arctic':
             return Arctic(config.arctic)
         elif store == 'influx':