+++ cryptostore/exceptions.py
@@ -1,8 +1,14 @@
 '''
-Copyright (C) 2018-2020  Bryant Moscon - bmoscon@gmail.com
+Copyright (C) 2018-2021  Bryant Moscon - bmoscon@gmail.com
 
 Please see the LICENSE file for the terms and conditions
 associated with this software.
 '''
+
+
 class InconsistentStorage(Exception):
+    pass
+
+
+class EngineWriteError(Exception):
     pass