+++ cryptostore/plugin/controller.py
@@ -1,5 +1,5 @@
 '''
-Copyright (C) 2018-2020  Bryant Moscon - bmoscon@gmail.com
+Copyright (C) 2018-2021  Bryant Moscon - bmoscon@gmail.com
 
 Please see the LICENSE file for the terms and conditions
 associated with this software.
@@ -26,7 +26,7 @@ class PluginController:
                 self.cfg = Config(file_name)
 
     def start(self):
-        if 'plugins' in self.cfg:
+        if self.cfg and 'plugins' in self.cfg:
             if self.cfg and isinstance(self.cfg.plugins, dict):
                 for _, plugin in self.cfg.plugins.items():
                     module = plugin.module